CanvasHotkeys overlay
Created a Bold overlay for showing a set of Hotkeys over the Canvas. Has a Bold text outline - using Gdip - so that this will be shown even over lighter backgrounds.

Added gdip_all.ahk and gdip_ext.ahk to use gdip and some gdip extended functions for this bold text.